+++ src/main/java/electrodynamics/common/item/gear/armor/types/composite/CompositeArmorItem.java
@@ -2,12 +2,16 @@ package electrodynamics.common.item.gear.armor.types.composite;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.function.Consumer;
 
 import electrodynamics.DeferredRegisters;
 import electrodynamics.api.References;
 import electrodynamics.api.capability.ceramicplate.CapabilityCeramicPlate;
 import electrodynamics.api.capability.ceramicplate.CapabilityCeramicPlateHolderProvider;
+import electrodynamics.client.ClientRegister;
+import electrodynamics.client.render.model.armor.types.ModelCompositeArmor;
 import net.minecraft.ChatFormatting;
+import net.minecraft.client.model.HumanoidModel;
 import net.minecraft.core.NonNullList;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.network.chat.Component;
@@ -17,6 +21,7 @@ import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ArmorItem;
 import net.minecraft.world.item.ArmorMaterial;
@@ -25,6 +30,9 @@ import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.level.Level;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.api.distmarker.OnlyIn;
+import net.minecraftforge.client.IItemRenderProperties;
 import net.minecraftforge.common.capabilities.ICapabilityProvider;
 
 public class CompositeArmorItem extends ArmorItem {
@@ -33,6 +41,23 @@ public class CompositeArmorItem extends ArmorItem {
 
     public CompositeArmorItem(ArmorMaterial materialIn, EquipmentSlot slot) {
 	super(materialIn, slot, new Item.Properties().stacksTo(1).tab(References.CORETAB).fireResistant().setNoRepair());
+    }
+    
+    @Override
+    @OnlyIn(Dist.CLIENT)
+    public void initializeClient(Consumer<IItemRenderProperties> consumer) {
+    	consumer.accept(new IItemRenderProperties() {
+    		@Override
+    		public <A extends HumanoidModel<?>> A getArmorModel(LivingEntity entity, ItemStack stack, EquipmentSlot slot, A properties) {
+    			ModelCompositeArmor<LivingEntity> model = new ModelCompositeArmor<>(ClientRegister.COMPOSITE_ARMOR_LAYER.bakeRoot(), slot);
+    			
+    			model.crouching = properties.crouching;
+    			model.riding = properties.riding;
+    			model.young = properties.young;
+    			
+    			return (A) model;
+    		}
+    	});
     }
 
     @Override
